backend/requirements.txt
fastapi
uvicorn
oracledb  # Remplace cx_Oracle par ça cx_Oracle
python-jose
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings